requirements.txt
requests>=2.31.0
requests-cache>=1.1.1
pymongo>=4.6.0
python-dateutil>=2.8.2
python-dotenv>=1.1.1

# Dev tools
black>=24.3.0
ruff>=0.3.0
pytest>=7.4.0
